% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_extreme_impact.R
\name{run_extreme_impact}
\alias{run_extreme_impact}
\title{run_extreme_impact}
\usage{
run_extreme_impact(inputsList = c(temp = NULL), agg_impacts = TRUE)
}
\arguments{
\item{inputsList}{}
}
\description{
run_extreme_impact
}
